tracing: 古いトレース機能を非推奨にし、新しいcloudtrace機能を追加
